Introduce sbroker_time for consistent time handling
sbroker_time handles the time in the native time unit of the VM and
provides conversion functions to/from milliseconds. In releases less
than 18.0 erlang:now/0 is used to tell the time and the native time
units is microseconds. For releases greater than, and including, 18.0
erlang:monotonic_time/0 is used and multi time warps will provide the
most accurate time.

Change sbroker and sregulator to use sbroker_time:

* Change the time unit to the native time unit, from milliseconds.
* Include the message queue sojourn time in the queue sojourn time.
* Infinity timeout calls return call time as last element in tuple.[1]
* Non-infinity timeout calls require timeout argument.[2]

[1] sregulator:drop/2 returns same as update/3 when response argument is
{drop, SojournTime}.

[2] cancel/2 -> cancel/3, ensure_dropped/2 -> ensure_dropped/2 and
done/2 -> done/3 with timeout as new third argument.