Add RelativeTime to go tuples
Add RelativeTime to go tuples to measure queue congestion. RelativeTime
is an estimation of the SojournTime if processing requests was instant.
A negative RelativeTime means that a request made RelativeTime earlier
would have succeeded without being queued.

In sbroker the RelativeTime is the difference between when the ask
request was sent and when the matching request was sent, in native time
units. Thus the difference is the potential sojourn time if using the
broker had no overhead.

The RelativeTime of the matched Pid is always -RelativeTime. A good
approximation of the matched Pid's sojourn time is
SojournTime - RelativeTime.

In sregulator the RelativeTime is the time spent in the internal queue
(sregulator_valve). As the regulator process decides when to dequeue an
item, load on the regulator may still have an effect. The same method
as sbroker can not be used because svalve callback modules may use
periods of time, rather than individual update events, to open the
valve.

The following formula can describe how relative time is calculated:

SojournTime = QueueDelay + ProcessingDelay
RelativeTime = QueueDelay - CounterPartyQueueDelay

QueueDelay is the delay due to waiting in queue for a counter party
request. ProcessingDelay is the delay due to processing the request and
all other requests while the request is queued. CounterPartyQueueDelay
is the delay for a counter party request waiting in a queue.

In regulator the CounerPartQueueDelay is unknown and assumed to be 0.
This assumption is incorrect.